% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_batch_monitor.R
\name{bn.batch.monitor}
\alias{bn.batch.monitor}
\title{A function to compute the batch monitor of a BN}
\usage{
bn.batch.monitor(df, child, parents, parent.values)
}
\arguments{
\item{df}{data in question}
}
\description{
A function to compute the batch monitor of a BN
}
\examples{
 cegb.prior <- get.ref.prior(df, cegb.struct, cuts, cegb.stage.key, cegb.stages)
 bn.batch.monitor()
}
\keyword{batch}
\keyword{cut}
