% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_node_monitors.R
\name{bn.uncondtnl.node.monitor}
\alias{bn.uncondtnl.node.monitor}
\title{A function to compute the node monitors of the BN}
\usage{
bn.uncondtnl.node.monitor(df, col_name = "Events", prior, n = 50,
  learn = FALSE)
}
\arguments{
\item{df}{data in question}

\item{n}{number of iterations}
}
\description{
A function to compute the node monitors of the BN
}
\examples{
###############################################################
}
\keyword{bn}
\keyword{monitor}
\keyword{node}
