% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_pa_child_monitor.R
\name{ceg.child.parent.monitor}
\alias{ceg.child.parent.monitor}
\title{A function to compute the node monitors of the BN}
\usage{
ceg.child.parent.monitor(df, target.stage, target.cut, condtnl.stage, stage.key,
  struct, stages, prior, n = 50, learn = FALSE)
}
\arguments{
\item{df}{data in question}

\item{stages}{}

\item{n}{number of iterations}

\item{learn}{binary to control learning}
}
\description{
A function to compute the node monitors of the BN
}
\examples{
###############################################################
}
\keyword{bn}
\keyword{monitor}
\keyword{node}
