% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceg_stage_monitors.R
\name{ceg.uncondtnl.stage.monitor}
\alias{ceg.uncondtnl.stage.monitor}
\title{A function to compute the node monitors of the BN}
\usage{
ceg.uncondtnl.stage.monitor(df, target.stage, target.cut, stages, stage.key,
  struct, n = 50, learn = FALSE)
}
\arguments{
\item{df}{data in question}

\item{stages}{}

\item{n}{number of iterations}

\item{learn}{binary to control learning}
}
\description{
A function to compute the node monitors of the BN
}
\examples{
###############################################################
}
\keyword{bn}
\keyword{monitor}
\keyword{node}
