% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostagekey.R
\name{find.cut}
\alias{find.cut}
\title{A function that can take the output from CEG.AHC() and put it into a list of data frames}
\usage{
find.cut(x)
}
\arguments{
\item{x}{not sure}
}
\description{
A function that can take the output from CEG.AHC() and put it into a list of data frames
}
\keyword{ceg,}
\keyword{structure}
