% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass.message.R
\name{get.edge.path.key}
\alias{get.edge.path.key}
\title{get edge key: dependency for messag passing algorithm}
\usage{
get.edge.path.key(stage.key)
}
\arguments{
\item{stage.key}{}
}
\description{
get edge key: dependency for messag passing algorithm
}
\examples{


}
\keyword{ceg}
\keyword{message}
\keyword{passing,}
