% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getprior.R
\name{get.ref.prior}
\alias{get.ref.prior}
\title{A function to set the reference prior on a CEG}
\usage{
get.ref.prior(df, struct, cuts, stage.key, stages)
}
\arguments{
\item{df}{data in question}

\item{stage.key}{which arrows come from which stages}

\item{stages}{input of stage names should be a character that ends in a numeric value}
}
\description{
A function to set the reference prior on a CEG
}
\examples{
 cegb.prior <- get.ref.prior(df, cegb.struct, cuts, cegb.stage.key, cegb.stages) 
 get.ref.prior()
}
\keyword{prior}
