% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_ceg.R
\name{render.ceg}
\alias{render.ceg}
\title{A function to automatically visualize the CEG}
\usage{
render.ceg(struct, cuts, stage.key, stages)
}
\arguments{
\item{stage.key}{which arrows come from which stages}

\item{stages}{input of stage names should be a character that ends in a numeric value}
}
\description{
This funciton takes all the possible path and divides by the number of florets in each stage to find the reference prior
 @param df data in question
}
\examples{
render.ceg()
}
\keyword{prior}
